% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_df_b}
\alias{example_df_b}
\title{Modified version of \code{datasets::mtcars} - version b}
\format{
A data frame with 9 rows and 9 variables:
\describe{
\item{car}{The rowname in the corresponding \code{datasets::mtcars} row}
\item{wt}{Weight (1000 lbs)}
\item{mpg}{Miles/(US) gallon}
\item{hp}{Gross horsepower}
\item{cyl}{Number of cylinders}
\item{disp}{Displacement (cu.in.)}
\item{carb}{Number of carburetors}
\item{drat}{Rear axle ratio}
\item{vs}{Engine (0 = V-shaped, 1 = straight)}
}
}
\source{
Sourced from the CRAN datasets package, with modified values. Originally from
Henderson and Velleman (1981), Building multiple regression models interactively. \emph{Biometrics}, \strong{37}, 391–411.
}
\usage{
example_df_b
}
\description{
A version of mtcars with some values altered and some rows/columns removed. Not for
informational purposes, used only to demonstrate the comparison of two slightly
different data frames. Since some values were altered at random, the values do
not necessarily reflect the true original values. The variables are as follows:
}
\keyword{datasets}
