% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-diff-rows.R
\name{not_equal}
\alias{not_equal}
\title{Test two vectors for not equal values}
\usage{
not_equal(col_a, col_b, allow_both_NA = TRUE)
}
\arguments{
\item{col_a, col_b}{Two equal length vectors, typically column values.}

\item{allow_both_NA}{If \code{TRUE} (default), an \code{NA} value in \code{col_a} and \code{col_b} is treated as equal.}
}
\description{
Test two vectors for not equal values
}
\keyword{internal}
