% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-versus.R
\name{print.tbl_versus}
\alias{print.tbl_versus}
\alias{tbl_format_header.tbl_versus}
\title{print tables with differing values}
\usage{
\method{print}{tbl_versus}(x, ...)

\method{tbl_format_header}{tbl_versus}(x, ...)
}
\description{
print tables with differing values
}
\keyword{internal}
