% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value-diffs.R
\name{value_diffs}
\alias{value_diffs}
\alias{value_diffs_stacked}
\alias{value_diffs_all}
\title{Show the differing values from a comparison}
\usage{
value_diffs(comparison, column)

value_diffs_stacked(comparison, column)

value_diffs_all(comparison)
}
\arguments{
\item{comparison}{The output of a \code{versus::compare()} call}

\item{column}{<\code{\link[=versus_tidy_select]{tidy-select}}>. The output will show the differing values
for the provided columns.}
}
\value{
\item{\code{value_diffs()}}{A data frame with one row for each element
of \code{col} found to be unequal between the input tables (
\code{table_a} and \code{table_b} from the original \code{compare()} call)
The output table has columns "val_a"/"val_b": the value of \code{col} in the
input tables, and the \code{by} columns for the identified rows in the
input tables.}

\item{\code{value_diffs_stacked()}, \code{value_diffs_all()}}{A data frame of
the \code{value_diffs()} output for the specified columns (or all columns if
\code{value_diffs_all()})
combined row-wise into a single table. To facilitate this combination into a
single table, the "val_a" and "val_b" columns are coerced to character.}
}
\description{
Show the differing values from a comparison
}
