% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versus.R
\docType{package}
\name{versus-package}
\alias{versus}
\alias{versus-package}
\title{versus: Compare Data Frames}
\description{
Compare two tables
}
\seealso{
Useful links:
\itemize{
  \item \url{https://eutwt.github.io/versus/}
  \item \url{https://github.com/eutwt/versus}
  \item Report bugs at \url{https://github.com/eutwt/versus/issues}
}

}
\author{
\strong{Maintainer}: Ryan Dickerson \email{fresh.tent5866@fastmail.com} [copyright holder]

Other contributors:
\itemize{
  \item Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID}) [contributor]
}

}
\keyword{internal}
