% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice-diffs.R
\name{slice_diffs}
\alias{slice_diffs}
\title{Show rows which have differing values in a comparison}
\usage{
slice_diffs(table, comparison, column = everything())
}
\arguments{
\item{table}{A data frame. One of \code{table_a} or \code{table_b} used to
create \code{comparison}}

\item{comparison}{The output of \code{compare()}}

\item{column}{<\code{\link[=versus_tidy_select]{tidy-select}}>. A row will be in the output if
the comparison shows differing values for any columns matching this argument}
}
\value{
The input \code{table} is filtered to the rows for which
\code{comparison} shows differing values for one of the columns in the supplied
\code{column} argument
}
\description{
Show rows which have differing values in a comparison
}
\examples{
comp <- compare(example_df_a, example_df_b, by = car)
example_df_a |> slice_diffs(comp, c(disp, mpg))
}
